Moved http-specific code to http.go
