+++ lfs/http.go
@@ -380,3 +380,234 @@ func (self *HttpApiContext) UploadObject(o *ObjectResource, reader io.Reader) *W
 
 	return wErr
 }
+
+func (self *HttpApiContext) doHttpRequest(req *http.Request, creds Creds) (*http.Response, *WrappedError) {
+	res, err := DoHTTP(Config, req)
+
+	var wErr *WrappedError
+
+	if err != nil {
+		wErr = Errorf(err, "Error for %s %s", res.Request.Method, res.Request.URL)
+	} else {
+		if creds != nil {
+			saveCredentials(creds, res)
+		}
+
+		wErr = self.handleResponse(res)
+	}
+
+	if wErr != nil {
+		if res != nil {
+			self.setErrorResponseContext(wErr, res)
+		} else {
+			self.setErrorRequestContext(wErr, req)
+		}
+	}
+
+	return res, wErr
+}
+
+func (self *HttpApiContext) doApiRequestWithRedirects(req *http.Request, creds Creds, via []*http.Request) (*http.Response, *WrappedError) {
+	res, wErr := self.doHttpRequest(req, creds)
+	if wErr != nil {
+		return nil, wErr
+	}
+
+	if res.StatusCode == 307 {
+		redirectedReq, redirectedCreds, err := self.newClientRequest(req.Method, res.Header.Get("Location"))
+		if err != nil {
+			return nil, Errorf(err, err.Error())
+		}
+
+		via = append(via, req)
+		seeker, ok := req.Body.(io.Seeker)
+		if !ok {
+			return nil, Errorf(nil, "Request body needs to be an io.Seeker to handle redirects.")
+		}
+
+		if _, err := seeker.Seek(0, 0); err != nil {
+			return nil, Error(err)
+		}
+		redirectedReq.Body = req.Body
+		redirectedReq.ContentLength = req.ContentLength
+
+		if err = checkRedirect(redirectedReq, via); err != nil {
+			return nil, Errorf(err, err.Error())
+		}
+
+		return self.doApiRequestWithRedirects(redirectedReq, redirectedCreds, via)
+	}
+
+	return res, nil
+}
+
+func (self *HttpApiContext) doApiRequest(req *http.Request, creds Creds) (*http.Response, *ObjectResource, *WrappedError) {
+
+	via := make([]*http.Request, 0, 4)
+	res, wErr := self.doApiRequestWithRedirects(req, creds, via)
+	if wErr != nil {
+		return nil, nil, wErr
+	}
+
+	obj := &ObjectResource{}
+	wErr = self.decodeApiResponse(res, obj)
+
+	if wErr != nil {
+		self.setErrorResponseContext(wErr, res)
+		return nil, nil, wErr
+	}
+
+	return res, obj, nil
+}
+
+func (self *HttpApiContext) doApiBatchRequest(req *http.Request, creds Creds) (*http.Response, []*ObjectResource, *WrappedError) {
+	via := make([]*http.Request, 0, 4)
+	res, wErr := self.doApiRequestWithRedirects(req, creds, via)
+	if wErr != nil {
+		return res, nil, wErr
+	}
+
+	var objs map[string][]*ObjectResource
+	wErr = self.decodeApiResponse(res, &objs)
+
+	if wErr != nil {
+		self.setErrorResponseContext(wErr, res)
+	}
+
+	return res, objs["objects"], wErr
+}
+
+func (self *HttpApiContext) handleResponse(res *http.Response) *WrappedError {
+	if res.StatusCode < 400 {
+		return nil
+	}
+
+	defer func() {
+		io.Copy(ioutil.Discard, res.Body)
+		res.Body.Close()
+	}()
+
+	cliErr := &ClientError{}
+	wErr := self.decodeApiResponse(res, cliErr)
+	if wErr == nil {
+		if len(cliErr.Message) == 0 {
+			wErr = self.defaultError(res)
+		} else {
+			wErr = Error(cliErr)
+		}
+	}
+
+	wErr.Panic = res.StatusCode > 499 && res.StatusCode != 501 && res.StatusCode != 509
+	return wErr
+}
+
+func (self *HttpApiContext) decodeApiResponse(res *http.Response, obj interface{}) *WrappedError {
+	ctype := res.Header.Get("Content-Type")
+	if !(lfsMediaTypeRE.MatchString(ctype) || jsonMediaTypeRE.MatchString(ctype)) {
+		return nil
+	}
+
+	err := json.NewDecoder(res.Body).Decode(obj)
+	io.Copy(ioutil.Discard, res.Body)
+	res.Body.Close()
+
+	if err != nil {
+		return Errorf(err, "Unable to parse HTTP response for %s %s", res.Request.Method, res.Request.URL)
+	}
+
+	return nil
+}
+
+func (self *HttpApiContext) defaultError(res *http.Response) *WrappedError {
+	var msgFmt string
+
+	if f, ok := defaultErrors[res.StatusCode]; ok {
+		msgFmt = f
+	} else if res.StatusCode < 500 {
+		msgFmt = defaultErrors[400] + fmt.Sprintf(" from HTTP %d", res.StatusCode)
+	} else {
+		msgFmt = defaultErrors[500] + fmt.Sprintf(" from HTTP %d", res.StatusCode)
+	}
+
+	return Error(fmt.Errorf(msgFmt, res.Request.URL))
+}
+
+func (self *HttpApiContext) newApiRequest(method, oid string) (*http.Request, Creds, error) {
+	endpoint := self.Endpoint()
+	objectOid := oid
+	operation := "download"
+	if method == "POST" {
+		if oid != "batch" {
+			objectOid = ""
+			operation = "upload"
+		}
+	}
+
+	res, err := sshAuthenticate(endpoint, operation, oid)
+	if err != nil {
+		tracerx.Printf("ssh: attempted with %s.  Error: %s",
+			endpoint.SshUserAndHost, err.Error(),
+		)
+	}
+
+	if len(res.Href) > 0 {
+		endpoint.Url = res.Href
+	}
+
+	u, err := ObjectUrl(endpoint, objectOid)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	req, creds, err := self.newClientRequest(method, u.String())
+	if err != nil {
+		return nil, nil, err
+	}
+
+	req.Header.Set("Accept", mediaType)
+	if res.Header != nil {
+		for key, value := range res.Header {
+			req.Header.Set(key, value)
+		}
+	}
+
+	return req, creds, nil
+}
+
+func (self *HttpApiContext) newClientRequest(method, rawurl string) (*http.Request, Creds, error) {
+	req, err := http.NewRequest(method, rawurl, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	req.Header.Set("User-Agent", UserAgent)
+	creds, err := getCreds(req)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return req, creds, nil
+}
+
+func (self *HttpApiContext) setErrorResponseContext(err *WrappedError, res *http.Response) {
+	err.Set("Status", res.Status)
+	self.setErrorHeaderContext(err, "Request", res.Header)
+	self.setErrorRequestContext(err, res.Request)
+}
+
+func (self *HttpApiContext) setErrorRequestContext(err *WrappedError, req *http.Request) {
+	err.Set("Endpoint", Config.Endpoint().Url)
+	err.Set("URL", fmt.Sprintf("%s %s", req.Method, req.URL.String()))
+	self.setErrorHeaderContext(err, "Response", req.Header)
+}
+
+func (self *HttpApiContext) setErrorHeaderContext(err *WrappedError, prefix string, head http.Header) {
+	for key, _ := range head {
+		contextKey := fmt.Sprintf("%s:%s", prefix, key)
+		if _, skip := hiddenHeaders[key]; skip {
+			err.Set(contextKey, "--")
+		} else {
+			err.Set(contextKey, head.Get(key))
+		}
+	}
+}
